fb_login_project/requirements.txt
selenium
Pyaml
webdriver-manager
